Replace command line interface with GUI
Implement ncurses-based GUI for Boggle game. Game supports timer,
scoring, and displaying user prompts, a visual Boggle board, time
and running score. Game begins with a menu that allows user to display
game rules or begin a game.

ncurses was chosen to allow for GUI based on multiple windows which may
be manipulated and updated asynchronously while still being able to run
in a terminal.

Fixed bug with handling of 'QU' boggle cube and scoring of words of
length 6.